requirements.txt
openai>=1.0.0
PyMuPDF>=1.22.0
python-docx>=0.8.11
python-dotenv>=1.0.0
# OCR opcional (si está instalado Tesseract/Poppler)
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=9.5.0
